src/python/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
selenium==4.21.0
webdriver-manager==4.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
Pillow==10.3.0
python-multipart==0.0.9
